feat: 仅针对脚本iframe没有使用pagehide或unload时添加此功能
